+++ src/main/java/alfio/controller/api/admin/ExtensionApiController.java
@@ -20,6 +20,8 @@ package alfio.controller.api.admin;
 import alfio.controller.api.support.PageAndContent;
 import alfio.extension.Extension;
 import alfio.extension.ExtensionService;
+import alfio.manager.AccessService;
+import alfio.manager.EventManager;
 import alfio.manager.user.UserManager;
 import alfio.model.EventAndOrganizationId;
 import alfio.model.ExtensionLog;
@@ -27,13 +29,9 @@ import alfio.model.ExtensionSupport;
 import alfio.model.ExtensionSupport.ExtensionMetadataValue;
 import alfio.model.ExtensionSupport.ExtensionParameterMetadataAndValue;
 import alfio.model.user.Organization;
-import alfio.model.user.User;
-import alfio.repository.EventRepository;
-import alfio.repository.user.OrganizationRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.ResponseEntity;
@@ -69,13 +67,13 @@ public class ExtensionApiController {
 
     private final ExtensionService extensionService;
     private final UserManager userManager;
-    private final OrganizationRepository organizationRepository;
-    private final EventRepository eventRepository;
+    private final EventManager eventManager;
+    private final AccessService accessService;
 
 
     @GetMapping("")
     public List<ExtensionSupport> listAll(Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         return extensionService.listAll();
     }
 
@@ -96,7 +94,7 @@ public class ExtensionApiController {
 
     private ResponseEntity<SerializablePair<Boolean, String>> createOrUpdate(String previousPath, String previousName, Extension script, Principal principal) {
         try {
-            ensureAdmin(principal);
+            accessService.ensureAdmin(principal);
             extensionService.createOrUpdate(previousPath, previousName, script);
             return ResponseEntity.ok(SerializablePair.of(true, null));
         } catch (Throwable t) {
@@ -107,20 +105,20 @@ public class ExtensionApiController {
 
     @GetMapping("{path}/{name}")
     public ResponseEntity<ExtensionSupport> loadSingle(@PathVariable("path") String path, @PathVariable("name") String name, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         return extensionService.getSingle(URLDecoder.decode(path, StandardCharsets.UTF_8), name).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
 
     @DeleteMapping(value = "{path}/{name}")
     public void delete(@PathVariable("path") String path, @PathVariable("name") String name, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         extensionService.delete(path, name);
     }
 
     @PostMapping(value = "/{path}/{name}/toggle/{enable}")
     public void toggle(@PathVariable("path") String path, @PathVariable("name") String name, @PathVariable("enable") boolean enable, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         extensionService.toggle(path, name, enable);
     }
 
@@ -137,38 +135,37 @@ public class ExtensionApiController {
 
     @PostMapping("/setting/system/bulk-update")
     public void bulkUpdateSystem(@RequestBody List<ExtensionMetadataValue> toUpdate, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         ensureIdsArePresent(toUpdate, extensionService.getConfigurationParametersFor("-", "-%", "SYSTEM"));
         extensionService.bulkUpdateSystemSettings(toUpdate);
     }
 
     @DeleteMapping("/setting/system/{id}")
     public void deleteSystemSettingValue(@PathVariable("id") int id, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         extensionService.deleteSettingValue(id, "-");
     }
 
     @GetMapping("/setting/organization/{orgId}")
     public Map<Integer, List<ExtensionParameterMetadataAndValue>> getParametersFor(@PathVariable("orgId") int orgId, Principal principal) {
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
+        accessService.checkOrganizationOwnership(principal, orgId);
+        Organization org = userManager.findOrganizationById(orgId, principal.getName());
         return extensionService.getConfigurationParametersFor("-" + org.getId(), "-" + org.getId()+"-%", "ORGANIZATION")
             .stream().collect(Collectors.groupingBy(ExtensionParameterMetadataAndValue::getExtensionId));
     }
 
     @PostMapping("/setting/organization/{orgId}/bulk-update")
     public void bulkUpdateOrganization(@PathVariable("orgId") int orgId, @RequestBody List<ExtensionMetadataValue> toUpdate, Principal principal) {
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
+        accessService.checkOrganizationOwnership(principal, orgId);
+        Organization org = userManager.findOrganizationById(orgId, principal.getName());
         ensureIdsArePresent(toUpdate, extensionService.getConfigurationParametersFor("-" + org.getId(), "-" + org.getId()+"-%", "ORGANIZATION"));
         extensionService.bulkUpdateOrganizationSettings(org, toUpdate);
     }
 
     @DeleteMapping("/setting/organization/{orgId}/{id}")
     public void deleteOrganizationSettingValue(@PathVariable("orgId") int orgId, @PathVariable("id") int id, Principal principal) {
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
-        extensionService.deleteSettingValue(id, "-" + org.getId());
+        accessService.checkOrganizationOwnership(principal, orgId);
+        extensionService.deleteSettingValue(id, "-" + orgId);
     }
 
     @GetMapping("/setting/organization/{orgId}/event/{shortName}")
@@ -176,10 +173,8 @@ public class ExtensionApiController {
                                                                      @PathVariable("shortName") String eventShortName,
                                                                      Principal principal) {
 
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
-        var event = eventRepository.findOptionalEventAndOrganizationIdByShortName(eventShortName).orElseThrow(IllegalStateException::new);
-        ensureEventInOrganization(org, event);
+        accessService.checkEventOwnership(principal, eventShortName, orgId);
+        var event = eventManager.getOptionalEventAndOrganizationIdByName(eventShortName, principal.getName()).orElseThrow(IllegalStateException::new);
         String pattern = generatePatternFrom(event);
         return extensionService.getConfigurationParametersFor(pattern, pattern,"EVENT")
             .stream().collect(Collectors.groupingBy(ExtensionParameterMetadataAndValue::getExtensionId));
@@ -188,10 +183,9 @@ public class ExtensionApiController {
     @PostMapping("/setting/organization/{orgId}/event/{shortName}/bulk-update")
     public void bulkUpdateEvent(@PathVariable("orgId") int orgId, @PathVariable("shortName") String eventShortName,
                                 @RequestBody List<ExtensionMetadataValue> toUpdate, Principal principal) {
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
-        var event = eventRepository.findOptionalEventAndOrganizationIdByShortName(eventShortName).orElseThrow(IllegalStateException::new);
-        ensureEventInOrganization(org, event);
+        accessService.checkEventOwnership(principal, eventShortName, orgId);
+        Organization org = userManager.findOrganizationById(orgId, principal.getName());
+        var event = eventManager.getOptionalEventAndOrganizationIdByName(eventShortName, principal.getName()).orElseThrow(IllegalStateException::new);
         String pattern = generatePatternFrom(event);
         ensureIdsArePresent(toUpdate, extensionService.getConfigurationParametersFor(pattern, pattern, "EVENT"));
         extensionService.bulkUpdateEventSettings(org, event, toUpdate);
@@ -199,10 +193,8 @@ public class ExtensionApiController {
 
     @DeleteMapping("/setting/organization/{orgId}/event/{shortName}/{id}")
     public void deleteEventSettingValue(@PathVariable("orgId") int orgId, @PathVariable("shortName") String eventShortName, @PathVariable("id") int id, Principal principal) {
-        Organization org = organizationRepository.getById(orgId);
-        ensureOrganization(principal, org);
-        var event = eventRepository.findOptionalEventAndOrganizationIdByShortName(eventShortName).orElseThrow(IllegalStateException::new);
-        ensureEventInOrganization(org, event);
+        accessService.checkEventOwnership(principal, eventShortName, orgId);
+        var event = eventManager.getOptionalEventAndOrganizationIdByName(eventShortName, principal.getName()).orElseThrow(IllegalStateException::new);
         extensionService.deleteSettingValue(id, generatePatternFrom(event));
     }
 
@@ -226,22 +218,9 @@ public class ExtensionApiController {
                                                      @RequestParam(required = false, name = "name") String name,
                                                      @RequestParam(required = false, name = "type") ExtensionLog.Type type,
                                                      @RequestParam(required = false, name = "page", defaultValue = "0") Integer page, Principal principal) {
-        ensureAdmin(principal);
+        accessService.ensureAdmin(principal);
         final int pageSize = 50;
         Pair<List<ExtensionLog>, Integer> res = extensionService.getLog(StringUtils.trimToNull(path), StringUtils.trimToNull(name), type, pageSize, (page == null ? 0 : page) * pageSize);
         return new PageAndContent<>(res.getLeft(), res.getRight());
-    }
-
-    private void ensureAdmin(Principal principal) {
-        Validate.isTrue(userManager.isAdmin(userManager.findUserByUsername(principal.getName())));
-    }
-
-    private void ensureOrganization(Principal principal, Organization organization) {
-        User user = userManager.findUserByUsername(principal.getName());
-        Validate.isTrue(userManager.isOwnerOfOrganization(user, organization.getId()));
-    }
-
-    private static void ensureEventInOrganization(Organization organization, EventAndOrganizationId event) {
-        Validate.isTrue(organization.getId() == event.getOrganizationId());
     }
 }
